feat: click to select word
Based off of  HugoLiconV's code. Removed sortable code and add onclick code.

I also changed the cursor when hovering over a button so the user knows to click.
It was a bit odd with the cursor never changing when using padding so i switched
to margin and increased the horizontal margin a bit.